Refactored the skillsMenu by adding an external function to construct it
